+++ Common/src/main/java/fuzs/puzzleslib/api/client/init/v1/ModelLayerFactory.java
@@ -5,6 +5,8 @@ import net.minecraft.client.model.geom.ModelLayerLocation;
 
 /**
  * A helper class for creating {@link ModelLayerLocation ModelLayerLocations} with a provided namespace.
+ * <p>
+ * TODO add ModelLayer suffix to method names to make the interface more usable in other context implementations
  */
 @FunctionalInterface
 public interface ModelLayerFactory {