+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v21.3.18-1.21.3] - 2025-01-08
 ### Added
 - Add `ArmorMaterialBuilder`
+- Add `RegistryManager::registerRecipeBookCategory`
 - Add a few new transitive access wideners
 ### Fixed
 - Avoid unnecessarily high memory usage from holding on to top level model locations after model events have completed